+++ src/renderer/components/Main.tsx
@@ -4,14 +4,14 @@ import Config from './Config';
 import InstalledMaps from './InstalledMaps';
 import MapTable from './MapTable';
 import MapTableHelp from './MapTableHelp';
-import OpenFolderButton from './OpenFolderButton';
 import AlertDialog from './shared/AlertDialog';
+import StormMapGenerator from './StormMapGenerator';
 
 export default function Main() {
   const { state } = useContext(GlobalContext);
 
   return (
-    <>
+    <div style={{ paddingTop: '80px' }}>
       {/* <h1>{state?.settings?.appName}</h1>
       <p>
         <span>Thank you~~</span>
@@ -27,12 +27,13 @@ export default function Main() {
                 <InstalledMaps />
               </>
             )}
+          {state?.settings?.showStormMapGenerator && <StormMapGenerator />}
           <h3>Map Installation:</h3>
           <MapTableHelp />
           <MapTable />
         </>
       )}
       <AlertDialog />
-    </>
+    </div>
   );
 }